fix 64bit problem

+++ trunk/external/CImg.h
@@ -12422,7 +12422,7 @@ namespace cimg_library {
         for (char *s = se2; s>ss; --s) if (*s==';' && level[s-expr._data]==clevel) { compile(ss,s); _cimg_mp_return(compile(s+1,se)); }
         for (char *s = ss1, *ps = ss, *ns = ss2; s<se1; ++s, ++ps, ++ns)
            if (*s=='=' && *ns!='=' && *ps!='=' && *ps!='>' && *ps!='<' && *ps!='!' && level[s-expr._data]==clevel) {
-             CImg<charT> variable_name((unsigned int)ss,(unsigned int)(s-ss+1)); variable_name.back() = 0;
+             CImg<charT> variable_name((size_t)ss,(size_t)(s-ss+1)); variable_name.back() = 0;
              bool is_valid_name = true;
              if ((*ss>='0' && *ss<='9') ||
                  (s==ss+1 && (*ss=='x' || *ss=='y' || *ss=='z' || *ss=='c' ||
@@ -12558,7 +12558,7 @@ namespace cimg_library {
         }
 
         // No known item found, assuming this is an already initialize variable.
-        CImg<charT> variable_name((unsigned int)ss,(unsigned int)(se-ss+1)); variable_name.back() = 0;
+        CImg<charT> variable_name((unsigned long)ss,(unsigned long)(se-ss+1)); variable_name.back() = 0;
         for (unsigned int i = 0; i<mempos; ++i) if (label[i]._data && !std::strcmp(variable_name,label[i])) _cimg_mp_return(i);
         *se = saved_char;
         throw CImgArgumentException("CImg<%s>::%s() : Unknown item '%s', in expression '%s'.\n",
